791 fixes and remaining updates for contrib to 100% coverage - broken out of SN2026
