+++ nautobot_ssot/integrations/solarwinds/diffsync/models/base.py
@@ -9,8 +9,8 @@ from nautobot.extras.models import Role
 from nautobot.ipam.models import IPAddress, IPAddressToInterface, Prefix
 
 from nautobot_ssot.contrib.model import NautobotModel
+from nautobot_ssot.contrib.typeddicts import ContentTypeDict
 from nautobot_ssot.contrib.types import CustomFieldAnnotation
-from nautobot_ssot.tests.contrib_base_classes import ContentTypeDict
 
 
 class LocationModel(NautobotModel):
