+++ nautobot_ssot/tests/contrib/test_object_metadata.py
