+++ nautobot_ssot/tests/contrib/test_sorting.py
@@ -1,8 +1,8 @@
 """Unit tests for contrib sorting."""
 
+import sys
 from typing import Annotated, List, Optional
-from unittest import skip
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
 
 from nautobot.apps.testing import TestCase
 from nautobot.extras.models import Tag
@@ -12,6 +12,7 @@ from typing_extensions import TypedDict, get_type_hints
 from nautobot_ssot.contrib import NautobotAdapter, NautobotModel
 from nautobot_ssot.contrib.sorting import (
     _is_sortable_field,
+    _sort_dict_attr,
     get_sort_key_from_typed_dict,
     get_sortable_fields_from_model,
     sort_relationships,
@@ -69,7 +70,6 @@ class TestAdapter(NautobotAdapter):
 ##############
 
 
-@skip("Sorting disabled.")
 class TestCaseIsSortableFieldFunction(TestCase):
     """Tests for `_is_sortable_field` function."""
 
@@ -87,7 +87,6 @@ class TestCaseIsSortableFieldFunction(TestCase):
         self.assertFalse(test)
 
 
-@skip("Sorting disabled.")
 class TestCaseGetSortedAttributesFromModel(TestCase):
     """Tests for `get_sortable_fields_from_model` function."""
 
@@ -100,7 +99,6 @@ class TestCaseGetSortedAttributesFromModel(TestCase):
         self.assertTrue(len(result) == 0)
 
 
-@skip("Sorting disabled.")
 class TestCaseSortRelationships(TestCase):
     """Tests for `sort_relationships` function."""
 
@@ -147,7 +145,6 @@ class TestCaseSortRelationships(TestCase):
         self.assertTrue(self.target.get_all("tenant")[0].tags[0]["name"] == "A Tag")
 
 
-@skip("Sorting disabled.")
 class TestGetSortKeyFromTypedDict(TestCase):
     """Unit tests for `get_sort_key_from_typed_dict` function."""
 
@@ -166,3 +163,75 @@ class TestGetSortKeyFromTypedDict(TestCase):
     def test_typed_dict_without_sort_key(self):
         """Test a typed dict without sort key specified."""
         self.assertIsNone(get_sort_key_from_typed_dict(BasicTagDict))
+
+
+class NautobotTenantUnsortableTags(NautobotModel):
+    """Tenant model whose list attribute has no SortKey, so it is not sortable."""
+
+    _model = Tenant
+    _modelname = "tenant"
+    _identifiers = ("name",)
+    _attributes = ("tags",)
+
+    name: str
+    tags: List[BasicTagDict] = []
+
+
+class NoSortFieldsAdapter(NautobotAdapter):
+    """Adapter whose top-level model exposes no sortable fields."""
+
+    top_level = ("tag",)
+    tag = BasicNautobotTag
+
+
+class SortingEdgeCaseTests(TestCase):
+    """Cover the remaining edge branches in contrib.sorting."""
+
+    def test_is_sortable_field_non_typing_returns_false(self):
+        """A hint without a name attribute is not sortable (hits the AttributeError guard)."""
+        self.assertFalse(_is_sortable_field(object()))
+
+    def test_is_sortable_field_python_3_9_branch(self):
+        """Force the Python <= 3.9 code path that reads `_name`."""
+        with patch.object(sys, "version_info", (3, 9, 0, "final", 0)):
+            self.assertFalse(_is_sortable_field(object()))
+
+    def test_sortable_field_without_sort_key_is_skipped(self):
+        """A list field whose TypedDict has no SortKey is not treated as sortable."""
+        self.assertEqual(get_sortable_fields_from_model(NautobotTenantUnsortableTags), {})
+
+    def test_sort_dict_attr_without_key_sorts_plainly(self):
+        """`_sort_dict_attr` with no key sorts the raw values."""
+
+        class _Holder:
+            values = ["b", "a", "c"]
+
+        holder = _Holder()
+        result = _sort_dict_attr(holder, "values", None)
+        self.assertEqual(result.values, ["a", "b", "c"])
+
+    def test_sort_relationships_noop_without_adapters(self):
+        """`sort_relationships` returns early when either adapter is missing."""
+        self.assertIsNone(sort_relationships(None, MagicMock()))
+
+    def test_sort_relationships_model_without_sortable_fields(self):
+        """A top-level model with no sortable fields is skipped without error."""
+        source = NoSortFieldsAdapter(job=MagicMock())
+        target = NoSortFieldsAdapter(job=MagicMock())
+        # Adapters must be non-empty, otherwise sort_relationships returns early.
+        for adapter in (source, target):
+            adapter.add(BasicNautobotTag(name="tag1"))
+        sort_relationships(source, target)  # exercises the "no sortable fields" continue
+
+    def test_sort_relationships_skips_falsy_model(self):
+        """A top-level attribute resolving to a falsy model is skipped.
+
+        diffsync forbids a falsy model at class-definition time, so simulate it by
+        overriding the resolved attribute on the instance.
+        """
+        source = NoSortFieldsAdapter(job=MagicMock())
+        target = NoSortFieldsAdapter(job=MagicMock())
+        for adapter in (source, target):
+            adapter.add(BasicNautobotTag(name="tag1"))
+        target.tag = None
+        sort_relationships(source, target)  # getattr(target, "tag") is None -> continue
